Move all CommunicationsManager related SQL from out of SKCase and into CommunicationsManager
